Align some example
