--- yarel-examples/basic/src/base/baseTest.java
@@ -1,11 +1,0 @@
-package base;
-import yarelcore.*;
-import java.util.Arrays;
-
-public class BaseTest
-{
-	public static void main(String[] args)
-	{
-		//Scrivere i test qui
-	}
-}
