+++ yarel-examples/basic/src/compound/CompoundTest.java
@@ -1,0 +1,11 @@
+package compound;
+import yarelcore.*;
+import java.util.Arrays;
+
+public class CompoundTest
+{
+	public static void main(String[] args)
+	{
+		//Scrivere i test qui
+	}
+}
