+++ yarel-examples/basic/src/compound/Decrement.java
@@ -2,11 +2,11 @@ package compound;
 import java.util.Arrays;
 import java.lang.Math;
 import yarelcore.*;	
-public class InvIncrement implements RPP {
-    public InvIncrement() { }
+public class Decrement implements RPP {
+    public Decrement() { }
     // Iteration start
     RPP function = new RPP() {
-    	private RPP f = new InvInc();
+    	private RPP f = new Dec();
     	private final int a = f.getA();
     	public int[] b(int[] x) {
     		return this.f.b(x);
