+++ yarel-examples/basic/src/compound/InvDecrement.java
