+++ yarel-examples/basic/src/compound/SeqComposition.java
@@ -2,10 +2,10 @@ package compound;
 import java.util.Arrays;
 import java.lang.Math;
 import yarelcore.*;	
-public class InvSeqComposition implements RPP {
-    public InvSeqComposition() { }
+public class SeqComposition implements RPP {
+    public SeqComposition() { }
     RPP l = new RPP() {
-    	RPP function = new InvIncrement();
+    	RPP function = new Increment();
     	private final int a = function.getA();
     	public int[] b(int[] x) { 
     		  	return this.function.b(x);
@@ -13,7 +13,7 @@ public class InvSeqComposition implements RPP {
     	 public int getA() { return this.a; }
     };
     RPP r = new RPP() {
-    	RPP function = new InvDecrement();
+    	RPP function = new Decrement();
     	private final int a = function.getA();
     	public int[] b(int[] x) { 
     		  	return this.function.b(x);
@@ -22,7 +22,7 @@ public class InvSeqComposition implements RPP {
     };
     private final int a = l.getA();
     public int[] b(int[] x) { // Implements a serial composition.
-    	return this.l.b(this.r.b(x));
+    	return this.r.b(this.l.b(x));
     }
     public int getA() { return this.a; }
 }
