--- yarel-examples/basic/src/compound/compoundTest.java
@@ -1,11 +1,0 @@
-package compound;
-import yarelcore.*;
-import java.util.Arrays;
-
-public class CompoundTest
-{
-	public static void main(String[] args)
-	{
-		//Scrivere i test qui
-	}
-}
