+++ yarel-examples/basic/src/esempiCompilazione.rl
@@ -1,0 +1,4 @@
+module esempiCompilazione {
+	dcl f: int
+	def f := inc
+}
