+++ yarel-examples/basic/src/esempiCompilazione/EsempiCompilazioneTest.java
@@ -1,0 +1,11 @@
+package esempiCompilazione;
+import yarelcore.*;
+import java.util.Arrays;
+
+public class EsempiCompilazioneTest
+{
+	public static void main(String[] args)
+	{
+		//Scrivere i test qui
+	}
+}
