+++ yarel-examples/basic/src/esempiCompilazione/F.java
@@ -1,10 +1,10 @@
-package innerFile;
+package esempiCompilazione;
 import java.util.Arrays;
 import java.lang.Math;
-import Yarelcore.*;
-public class R implements RPP {
-    public R() { }
-    private RPP f = new Id();
+import yarelcore.*;	
+public class F implements RPP {
+    public F() { }
+    private RPP f = new Inc();
     private final int a = f.getA();
     public int[] b(int[] x) {
     	return this.f.b(x);
