+++ yarel-examples/basic/src/esempiCompilazione/InvF.java
@@ -1,10 +1,10 @@
-package innerFile;
+package esempiCompilazione;
 import java.util.Arrays;
 import java.lang.Math;
-import Yarelcore.*;
-public class InvR implements RPP {
-    public InvR() { }
-    private RPP f = new InvId();
+import yarelcore.*;	
+public class InvF implements RPP {
+    public InvF() { }
+    private RPP f = new InvInc();
     private final int a = f.getA();
     public int[] b(int[] x) {
     	return this.f.b(x);
