--- yarel-examples/basic/src/innerFile/InnerFilePlayWith.java
@@ -1,12 +1,0 @@
-package innerFile;
-import Yarelcore.*;
-import java.util.Arrays;
-
-public class InnerFilePlayWith {
-	public static void main(String[] args) throws Exception {
-		 RPP rRPP = new innerFile.R();
-		 for(int i : rRPP.b(new int[] {5})) {
-		 		System.out.println(i);
-		 	}
-	}
-}
