+++ yarel-examples/basic/src/modules/ModulesTest.java
@@ -1,0 +1,11 @@
+package modules;
+import yarelcore.*;
+import java.util.Arrays;
+
+public class ModulesTest
+{
+	public static void main(String[] args)
+	{
+		//Scrivere i test qui
+	}
+}
