+++ yarel-examples/basic/src/modules/Permutation.java
@@ -1,11 +1,11 @@
-package compound;
+package modules;
 import java.util.Arrays;
 import java.lang.Math;
 import yarelcore.*;	
-public class InvSeqComposition implements RPP {
-    public InvSeqComposition() { }
+public class Permutation implements RPP {
+    public Permutation() { }
     RPP l = new RPP() {
-    	RPP function = new InvIncrement();
+    	RPP function = new compound.ParComposition();
     	private final int a = function.getA();
     	public int[] b(int[] x) { 
     		  	return this.function.b(x);
@@ -13,16 +13,19 @@ public class InvSeqComposition implements RPP {
     	 public int getA() { return this.a; }
     };
     RPP r = new RPP() {
-    	RPP function = new InvDecrement();
-    	private final int a = function.getA();
-    	public int[] b(int[] x) { 
-    		  	return this.function.b(x);
+    	private final int a = 4;
+    	public int[] b(int[] x) {
+    		int tmp=0;
+    		tmp = x[1]; 
+    		x[1] = x[2]; 
+    		x[2] = tmp; 
+    		return x;
     	}
-    	 public int getA() { return this.a; }
+    	public int getA() { return this.a; }
     };
     private final int a = l.getA();
     public int[] b(int[] x) { // Implements a serial composition.
-    	return this.l.b(this.r.b(x));
+    	return this.r.b(this.l.b(x));
     }
     public int getA() { return this.a; }
 }
