--- yarel-examples/basic/src/modules/modulesTest.java
@@ -1,11 +1,0 @@
-package modules;
-import yarelcore.*;
-import java.util.Arrays;
-
-public class ModulesTest
-{
-	public static void main(String[] args)
-	{
-		//Scrivere i test qui
-	}
-}
