+++ yarel-examples/basic/src/yarelcore/Id.java
@@ -1,0 +1,8 @@
+package yarelcore;
+public class Id implements RPP {
+	private final int a = 1;
+	public int[] b(int[] x) {
+		return x;
+	}
+	public int getA() { return this.a; }
+}
