+++ yarel-examples/basic/src/yarelcore/RPP.java
@@ -1,0 +1,5 @@
+package yarelcore;
+public interface RPP {
+	public int getA();
+	public int[] b(int[] x);
+}
