+++ yarel-examples/basic/src/yarelcore/WrongArityException.java
@@ -1,0 +1,9 @@
+package yarelcore;
+public class WrongArityException extends RuntimeException {
+    public WrongArityException(){
+        super();
+    }
+    public WrongArityException(String message){
+        super(message);
+    }
+}
